editing the image field in product admin

+++ products/views.py
@@ -75,6 +75,10 @@ def product_detail(request, product_id):
 @login_required
 def add_product(request):
     """ add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permissons for this action!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -97,6 +101,10 @@ def add_product(request):
 @login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permissons for this action!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -122,6 +130,10 @@ def edit_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permissons for this action!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'product deleted')
